+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 
 ### Fixed
 - Set connection runtime parameters, after switching to the main connection (cover uppercase SET)
+- Don't call `DirtyConnectionHook` after rollback
 
 ## [2.7.0] - 2022-09-08
 [2.7.0]: https://github.com/atlassian-labs/db-replica/compare/release-2.6.4...release-2.7.0